Changed speed slider, fixed metres display

+++ GlassFitUnity/Assets/David/PBRunnerController.cs
@@ -112,7 +112,7 @@ public class PBRunnerController : MonoBehaviour {
 	string SiDistance(double meters) {
 		string postfix = "m";
 		string final;
-		int value = (int)meters;
+		float value = (float)meters;
 		if (value > 1000) {
 			value = value/1000;
 			postfix = "km";
