+++ GlassFitUnity/Assets/Scripts/TrainController.cs
@@ -121,7 +121,7 @@ public class TrainController : MonoBehaviour {
 	string SiDistance(double meters) {
 		string postfix = "m";
 		string final;
-		int value = (int)meters;
+		float value = (float)meters;
 		if (value > 1000) {
 			value = value/1000;
 			postfix = "km";
