+++ GlassFitUnity/Assets/Scripts/ZombieController.cs
@@ -146,7 +146,7 @@ public class ZombieController : MonoBehaviour {
 	string SiDistance(double meters) {
 		string postfix = "m";
 		string final;
-		int value = (int)meters;
+		float value = (float)meters;
 		if (value > 1000) {
 			value = value/1000;
 			postfix = "km";
